refactor: split dashboard and settings pages
